### 2019-05-12 #51 @labarba
@AstrobioMike : You're missing the line `bibliography: paper.bib` in the header of your paper. Can you push this fix?
