### 2019-05-15 #53 @whedon
[REVIEW]: Happy Belly Bioinformatics: an open-source resource dedicated to helping biologists utilize bioinformatics
**Submitting author:** @AstrobioMike (<a href="http://orcid.org/0000-0001-7750-9145">Michael D. Lee</a>)
**Repository:** <a href="https://github.com/AstrobioMike/AstrobioMike.github.io" target ="_blank">https://github.com/AstrobioMike/AstrobioMike.github.io</a>
**Version:** v1.2.0
