### 2019-05-15 #53 @labarba
👋 @luizirber, @JasonJWilliamsNY — Thank you for agreeing to review this submission to JOSE. This is where it happens: you both have a checklist at the top of this issue to help you along. Feel free to post questions here and open issues on the submission repository, if needed. The version item is something we can check off in the end, as we will ask the authors to do a release after any revisions. Statement of need should be a section in the paper (the checklist has it also in the Documentation, but that's optional).

Please tick off your Conflict of Interest and Code of Conduct items ASAP, so we all see the review has started. Thanks!
