### 2019-05-15 #53 @labarba
Our [Reviewer guidelines](http://jose.theoj.org/about#reviewer_guidelines) cover briefly our expectations.
