### 2019-05-15 15a17b2
updating bash install page with conda note for now
