### 2019-05-16 b7cf1e6
minor updates
