+++ bash/index.md
@@ -7,7 +7,7 @@ permalink: /bash/index
 
 {% include _side_tab_bash.html %}
 
-Bash is the most common programming language in a Unix-like environment these days. More importantly, **bash is very likely the most foundational skillset you can learn for bioinformatics** (and much more). There is a [table of some quick terminology](/bash/bash_intro#some-terminology){:target="_blank"} for some help in what is meant by terms like "bash" and "Unix" on the [introduction to bash page](/bash/bash_intro#some-terminology){:target="_blank"}, which is definitely the place to start if you're completely new to this wonderful language. And if you would like some help getting a bash environment up and running on your computer, check out [this page](/bash/getting_bash_env){:target="_blank"} 🙂
+Bash is the most common programming language in a Unix-like environment these days. More importantly, **bash is very likely the most foundational skillset you can learn for bioinformatics** (and much more). There is a [table of some quick terminology](/bash/bash_intro_binder#some-terminology){:target="_blank"} for some help in what is meant by terms like "bash" and "Unix" on the [introduction to bash page](/bash/bash_intro_binder){:target="_blank"}, which is definitely the place to start if you're completely new to this wonderful world of Unix-like environments. And if you would like some help getting a bash environment up and running on your computer, check out [this page](/bash/getting_bash_env){:target="_blank"} 🙂
 
 
 <br>
@@ -29,13 +29,13 @@ Bash is the most common programming language in a Unix-like environment these da
 <br>
 # All bash pages
 * [Getting a bash environment](/bash/getting_bash_env)
-* [Intro to bash](/bash/bash_intro)
+* [Intro to bash](/bash/bash_intro_binder)
 * [Six glorious commands](/bash/six_commands)
 * [The wonderful world of loops](/bash/for_loops)
 * [Why is this all worth it?](/bash/why)
 * [What the heck is my PATH, and how do I get it to do what I want?](/bash/modifying_your_path)
 * [Help installing tools](/bash/installing_tools)
-* [Some useful one-liners](/bash/one_liners)
+<!--* [Some useful one-liners](/bash/one_liners)-->
 * [Downloading from NCBI](/bash/ncbi_eutils)
 * [Other great resources](/bash/other_resources)
 
