### 2019-06-13 4bc3f2b
updating de novo genome assembly page, anvio, and added conda instructions
