+++ genomics/de_novo_assembly.md
@@ -9,7 +9,7 @@ permalink: /genomics/de_novo_assembly
 
 {% include _side_tab_genomics.html %}
 
-Here we're going to run through some of the typical steps for taking a newly sequenced isolate genome from raw fastq files through to an assembled, curated genome you can then begin to explore. It's assumed you're already comfortable with working at the command line, if you're not yet, you should probably run through [*bash* basics](/bash/basics){:target="_blank"} first.  
+Here we're going to run through some of the typical steps for taking a newly sequenced isolate genome from raw fastq files through to an assembled, curated genome you can then begin to explore. It's assumed you're already somewhat familiar with working at the command line, if you're not yet, you should probably run through the [*bash* basics](/bash/bash_intro_binder){:target="_blank"} first 🙂  
 
 And before we get started here, a public service announcement:
 
@@ -24,15 +24,44 @@ Now that that's out of the way, let's get to it!
 ---
 <br>
 # Tools used here
-Throughout this process we'll be using a variety of tools that I've listed here, along with the particular versions I used while running things on either [my personal computer (MacOSX - Darwin) and/or on a server (Linux - Ubuntu)](https://www.quora.com/Whats-the-difference-between-Mac-OS-X-Darwin-OS-and-a-popular-Linux-distribution-like-Ubuntu-What-can-be-done-on-Darwin){:target="_blank"}. If you plan to follow along, you'll need to install these and have them working on your system. The links here go to their respective developer pages where you can usually find great help for installation, but I've also added example installations of all the tools used on this site on the [*bash* installing tools page](/bash/installing_tools){:target="_blank"}. So feel free to visit there for help also as you work through getting each of these installed.  
+Throughout this process we'll be using a variety of tools that I've listed below, along with the particular versions I used while putting this together. If you plan to follow along, you'll need to install these and have them working on your system. The links below go to their respective developer pages where you can usually find great help for installation, but you can also use the wonderful package manager [conda](https://conda.io/docs/){:target="_blank"} to install things. 
+
+To get conda up and running (which is very quick), you can follow the instructions to install miniconda (a light-weight version) for your appropriate system starting from [here](https://conda.io/en/latest/miniconda.html){:target="_blank"}, and then the conda installations below will do the trick.
 
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; • [FastQC v0.11.5](https://www.bioinformatics.babraham.ac.uk/projects/fastqc/){:target="_blank"}  
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; • [Trimmomatic v0.36](http://www.usadellab.org/cms/?page=trimmomatic){:target="_blank"}  
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; • [SPAdes v3.11.1](http://cab.spbu.ru/software/spades/){:target="_blank"}  
 &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; • [MegaHit v1.1.1](https://github.com/voutcn/megahit){:target="_blank"}  
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; • [QUAST v4.5](https://github.com/ablab/quast){:target="_blank"}  
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; • [QUAST v5.0.2](https://github.com/ablab/quast){:target="_blank"}  
[…]
