### 2019-06-18 d585baa
re-doing R basics with binder
