+++ R/basics.md
@@ -9,7 +9,7 @@ permalink: /R/basics
 
 {% include _side_tab_R.html %}
 
-This module is designed for those that are either completely new to R, or have some experience but maybe don't feel as solid about some of the fundamentals as they'd like. It will run through the very basics such as setting up your working environment, assigning variables, "indexing" (subsetting data), reading in and writing out data, and installing packages. Some relevant terminology is presented [here](/R/index) if you find yourself seeing some words that are unfamilar to you. This page is meant to be a quick-start to get you into and using the R environment. There is, of course, an incredible amount of functionality we won't be touching on here, so be sure to peruse the extensive [R intro documentation available here](https://cran.r-project.org/doc/manuals/r-release/R-intro.html) when you're starting to get your bearings and want to go further, and there are also some great courses available at [DataCamp](https://www.datacamp.com/) that are worth checking out after you have a solid foundation. 
+This module is designed for those that are either completely new to R, or have some experience but maybe don't feel as solid about some of the fundamentals as they'd like. It will run through the very basics such as setting up your working environment, assigning variables, "indexing" (subsetting data), reading in and writing out data, and installing packages. Some relevant terminology is presented [here](/R/index) if you find yourself seeing some words that are unfamilar to you. This page is meant to be a quick-start to get you into and using the R environment. There is, of course, an incredible amount of functionality we won't be touching on here, so be sure to peruse the extensive [R intro documentation available here](https://cran.r-project.org/doc/manuals/r-release/R-intro.html) when you're starting to get your bearings and want to go further 🙂 
 <br>
 <br>
 
@@ -17,22 +17,55 @@ This module is designed for those that are either completely new to R, or have s
 ---
 <br>
 
-# Installation
-It is possible your computer already has R, if you are unsure, you can check by opening a terminal window and typing `R`. If this launches R rather than giving an error message, you should be good to go (enter `q()` to exit the R environment). If you do not have R, you can download it from here for Mac: [https://cran.r-project.org/bin/macosx/](https://cran.r-project.org/bin/macosx/). And if you have a relatively newer Mac, you may also need to install XQuartz which you can get from here: [https://www.xquartz.org/](https://www.xquartz.org/). 
+# Accessing our R environment
+Before we get started, we need an R environment to work in. You can work on your own computer if you'd like, or you can work in a "Binder" that's been created for this page, see below. 
 
-Lastly, I highly, *highly*, **highly** recommend installing the free version of [RStudio](https://www.rstudio.com/) if you don't already have it. RStudio is an interface for R that not only makes everything you will do in R easier and more organized, but it's also invaluable for reproducibility of your analyses as it makes it second-nature to generate and save R scripts of everything you're doing, while you're doing it – which is very helpful when you want to look back and see what worked out of the 20 things you just tried 🙂. You can download an RStudio installer from [here](https://www.rstudio.com/products/rstudio/download/#download).  
+## On your computer
+It is possible your computer already has R, if you are unsure, you can check by opening a terminal window ([a unix-like terminal](/bash/getting_bash_env){:target="_blank"}) and typing `R`. If this launches R rather than giving an error message, you should be good to go (enter `q()` to exit the R environment). If you do not have R, you can download it from here for Mac: [https://cran.r-project.org/bin/macosx/](https://cran.r-project.org/bin/macosx/){:target="_blank"}. And if you have a relatively newer Mac, you may also need to install XQuartz which you can get from here: [https://www.xquartz.org/](https://www.xquartz.org/){:target="_blank"}. 
+
[…]
