### 2019-06-18 df59b0c
mod r installation page
