+++ R/installing_packages.md
@@ -9,7 +9,7 @@ permalink: /R/installing_packages
 
 {% include _side_tab_R.html %}
 
-Part of what makes R so valuable is that there is an enormous community of people developing software packages for it. People share bundles of code that perform specific tasks through what are known as "packages". Packages are maintained at the [Comprehensive R Archive Network (CRAN)](https://cran.r-project.org/web/packages/). To use a package, you simply need to install it and then load it. Here we'll cover the 3 main ways for installing packages.
+Part of what makes R so valuable is that there is an enormous community of people developing software packages for it. People share bundles of code that perform specific tasks through what are known as "packages". Packages are typically maintained at the [Comprehensive R Archive Network (CRAN)](https://cran.r-project.org/web/packages/){:target="_blank"} and/or at [Bioconductor](https://www.bioconductor.org/){:target="_blank"}. To use a package, you simply need to install it and then load it. Here we'll cover the 3 main ways for installing packages.
 <br>
 <br>
 
@@ -18,7 +18,7 @@ Part of what makes R so valuable is that there is an enormous community of peopl
 <br>
 
 ## install.packages()
-Most often, you will be able to install packages with the `install.packages()` function. For example, if you want to install the [ggplot2 package](https://cran.r-project.org/web/packages/ggplot2/index.html), you would simply enter:
+Most often, you will be able to install packages with the `install.packages()` function. For example, if you want to install the [ggplot2 package](https://cran.r-project.org/web/packages/ggplot2/index.html){:target="_blank"}, you would simply enter:
 
 ```R
 install.packages("ggplot2")
@@ -41,7 +41,7 @@ But do not despair! This is usually just a consequence of the package not having
 <br>
 
 ## biocLite()
-For example, searching for ["phyloseq bioconductor"](https://www.google.com/search?q=bioconductor+phyloseq+R&oq=bioconductor+phyloseq+R&aqs=chrome..69i57j69i60.5566j0j7&sourceid=chrome&ie=UTF-8) returns the [bioconductor homepage of the *phyloseq* package](http://bioconductor.org/packages/release/bioc/html/phyloseq.html) as the top hit. And when you head over there, there are directions to install via bioconductor:
+For example, searching for ["phyloseq bioconductor"](https://www.google.com/search?q=bioconductor+phyloseq+R&oq=bioconductor+phyloseq+R&aqs=chrome..69i57j69i60.5566j0j7&sourceid=chrome&ie=UTF-8) returns the [bioconductor homepage of the *phyloseq* package](http://bioconductor.org/packages/release/bioc/html/phyloseq.html){:target="_blank"} as the top hit. And when you head over there, there are directions to install via bioconductor:
 
 ```R
 source("https://bioconductor.org/biocLite.R")
@@ -62,7 +62,7 @@ And again, after installing, you'd need to load the library in order to access t
 ## install_github()
 Occasionally bioconductor may also not workout. At that point I usually turn to searching for the package on github as it may be hosted there, and if so it usually contains instructions for how to install it via devtools and the `install_github()` function. 
 
[…]
