+++ .github/ISSUE_TEMPLATE/new-topic-idea.md
@@ -1,0 +1,10 @@
+---
+name: New Topic Idea
+about: Suggest/Request a new topic to be covered on the site
+title: New Topic Idea
+labels: new topic idea
+assignees: ''
+
+---
+
+**What else do you think should be covered on the site?** (This can be a request for something you are or did have trouble with, or it can be a suggestion for something you're comfortable with but think will help others!)
