### 2019-06-18 8cd65f7
Update issue templates
