+++ .github/ISSUE_TEMPLATE/mistakes--bugs--or-typos.md
@@ -1,0 +1,16 @@
+---
+name: Mistakes, bugs, or typos
+about: Communicate any problems with the site
+title: Mistake, bug, or typo
+labels: problem
+assignees: ''
+
+---
+
+**Describe the mistake, bug, or typo here:**
+
+
+**Link to the page containing the issue:**
+
+
+**Suggested fix – if any, just reporting what's wrong is super-helpful either way :)**
