### 2019-06-19 790278b
updating bash intro, adding drop down challenge boxes
