+++ bash/bash_intro_binder.md
@@ -19,7 +19,7 @@ permalink: /bash/bash_intro_binder
 <br>
 
 # Important note!
-**Keep in mind, none of this is about memorization. It may seem counterintuitive, but the little details aren't important. What matters is starting to build a mental framework of the foundational rules and concepts. That equips us to figure out the things we need to do, when we need to do them!** 
+**Maybe the most important thing to keep in mind here is that this is all about *exposure*, not memorization or mastering anything. Don't worry about the details. At first we just need to starting building a mental framework of the foundational rules and concepts. That equips us to figure out the things we need to, when we need to do them 🙂 **
 
 <center>Here's a high-resolution timelapse of my ongoing journey:</center>
 <center><img src="{{ site.url }}/images/mike_philosophy2.png" title="Don't worry about every little detail!"></center>
@@ -37,10 +37,10 @@ First, here are some terms that are often used interchangeably.
 
 | Term     | What it is          |
 |:-------------:|------------------|
+| **`shell`** | what we use to talk to the computer; anything where you are pointing and clicking with a mouse is a **G**raphical **U**ser **I**nterface (**GUI**) shell; something with text only is a **C**ommand **L**ine **I**nterface (**CLI**) shell |  
 | **`command line`** | a text-based environment capable of taking input and providing output |  
 | **`Unix`** | a family of operating systems |  
 | **`bash`** | the most common programming language used at a Unix command-line |  
-| **`shell`** | what we use to talk to the computer; anything where you are pointing and clicking with a mouse is a **G**raphical **U**ser **I**nterface (**GUI**) shell; something with text only is a **C**ommand **L**ine **I**nterface (**CLI**) shell |  
 
 <br>
 
@@ -66,17 +66,18 @@ Before we get started, we need a terminal to work in. You can either work on you
 
 When that page finishes loading (it may take a couple minutes), you will see a screen like this (minus the blue arrows):
 
-<center><img src="{{ site.url }}/images/binder-app-launch.png"></center>
+<center><img src="../images/binder-app-launch.png"></center>
 <br>
 
 Now click the **folder icon** at the top-left (that the smaller blue arrow points to above) and then click the "**Terminal**" icon at the bottom, and we'll be in our appropriate command-line environment:
 
-<center><img src="{{ site.url }}/images/binder-terminal.png"></center>
+<center><img src="../images/binder-terminal.png"></center>
 <br>
 
 This is our "command line", where we will be typing all of our commands 🙂
 
 >**Note:** If you want to get a bash environment of your own going on your computer, you can see [this page](/bash/getting_bash_env){:target="_blank"} for some help, and can then follow this page working on your own computer.
+
 <br>
 
 ---
@@ -88,13 +89,27 @@ This is our "command line", where we will be typing all of our commands 🙂
 
 ```
 cd ~
-curl -LO https://github.com/AstrobioMike/AstrobioMike.github.io/raw/master/misc/bash_intro.tar.gz
-tar -xzvf bash_intro.tar.gz
-rm bash_intro.tar.gz
+curl -L -o bash_intro.tar.gz https://ndownloader.figshare.com/files/15537653
+tar -xzvf bash_intro.tar.gz && rm bash_intro.tar.gz
 cd bash_intro
 ```
 
 Don't forget to press enter to execute the last command (usually that doesn't copy over). If your system does not have `curl` installed, and you get an error message from the above, then it's probably best to work in the [binder environment](/bash/bash_intro_binder#accessing-our-command-line-environment) for now while getting used to things 🙂  
+
+<br>
+
+---
+<br>
+
+# A few foundational rules
+
[…]
