### 2019-06-19 b775918
adding more to bash resources page
