+++ bash/other_resources.md
@@ -8,6 +8,11 @@ permalink: /bash/other_resources
 
 # Here are some other great places to learn about bash!
 
-[Command-line bootcamp](http://rik.smith-unna.com/command_line_bootcamp/){:target="_blank"} setup by [Rik Smith-Unna](https://twitter.com/blahah404){:target="_blank"} provides an awesome self-contained bash environment alongside tutorials
+Google is always a great place to start if we have a specific problem we are hitting. It's most often going to be something others have already ran into and discussed 🙂
 
-[Explainshell.com](https://explainshell.com/){:target="_blank"} which allows you to paste in a command to have it broken down into individual parts
+Super helpful sites with excellent forums include [Biostars](https://www.biostars.org/){:target="_blank"} and [StackExchange](https://stackoverflow.com/){:target="_blank"}.
+
+[Command-line bootcamp](http://rik.smith-unna.com/command_line_bootcamp/){:target="_blank"} setup by [Rik Smith-Unna](https://twitter.com/blahah404){:target="_blank"} provides an awesome self-contained bash environment alongside tutorials.
+
+[Explainshell.com](https://explainshell.com/){:target="_blank"} allows pasting in a command to have it broken down into individual parts.
+
