### 2019-06-20 7eb93d0
updating R lessons, including binders, new screen capture images of binder environment added
