+++ R/basics.md
@@ -1,6 +1,6 @@
 ---
 layout: main
-title: R basics
+title: Introduction to R
 categories: [R, tutorial]
 permalink: /R/basics
 ---
@@ -8,6 +8,18 @@ permalink: /R/basics
 {% include _R_basics_toc.html %}
 
 {% include _side_tab_R.html %}
+
+> **Things covered here:**  
+*  RStudio value and layout
+*  How to check and set our working directory
+*  Basic calculations and setting variables
+*  What indexing is and how to do it
+*  Reading in and writing out data
+
+<br>
+
+# Important note!
+**Maybe the most important thing to keep in mind here is that this is all about *exposure*, not memorization or mastering anything. Don't worry about the details. At first we just need to starting building a mental framework of the foundational rules and concepts. That equips us to figure out the things we need to, when we need to do them** 🙂
 
 This module is designed for those that are either completely new to R, or have some experience but maybe don't feel as solid about some of the fundamentals as they'd like. It will run through the very basics such as setting up your working environment, assigning variables, "indexing" (subsetting data), reading in and writing out data. Some relevant terminology is presented [here](/R/index){:target="_blank"} if you find yourself seeing some words that are unfamilar to you. This page is meant to be a quick-start to get you into and using the R environment.
 
@@ -24,7 +36,7 @@ Part of what makes R so valuable and powerful are all of the open-source package
 Before we get started, we need an R environment to work in. You can work on your own computer if you'd like, or you can work in a "Binder" that's been created for this page, see below. 
 
 ## On your computer
-It is possible your computer already has R, if you are unsure, you can check by opening a terminal window ([a unix-like terminal](/bash/getting_bash_env){:target="_blank"}) and typing `R`. If this launches R rather than giving an error message, you should be good to go (enter `q()` to exit the R environment). If you do not have R, you can download it from here for Mac: [https://cran.r-project.org/bin/macosx/](https://cran.r-project.org/bin/macosx/){:target="_blank"}. And if you have a relatively newer Mac, you may also need to install XQuartz which you can get from here: [https://www.xquartz.org/](https://www.xquartz.org/){:target="_blank"}. 
+It is possible your computer already has R, if you are unsure, you can check by opening a terminal window ([a unix-like terminal](/bash/getting_bash_env){:target="_blank"}) and typing **`R`**. If this launches R rather than giving an error message, you should be good to go (enter **`q()`** to exit the R environment). If you do not have R, you can download it from here for Mac: [https://cran.r-project.org/bin/macosx/](https://cran.r-project.org/bin/macosx/){:target="_blank"}. And if you have a relatively newer Mac, you may also need to install XQuartz which you can get from here: [https://www.xquartz.org/](https://www.xquartz.org/){:target="_blank"}. 
 
 Lastly, I highly, *highly*, **highly** recommend installing the free version of [RStudio](https://www.rstudio.com/){:target="_blank"} if you don't already have it. RStudio is an interface for R that not only makes everything you will do in R easier and more organized, but it's also invaluable for reproducibility of your analyses as it makes it second-nature to generate and save R scripts of everything you're doing, while you're doing it – which is very helpful when you want to look back and see what worked out of the 20 things you just tried 🙂. You can download an RStudio installer from [here](https://www.rstudio.com/products/rstudio/download/#download){:target="_blank"}. See the next section for the typical layout of RStudio.  
 
@@ -48,11 +60,11 @@ If we then click the icon with the green plus sign at the top-left, and then "R 
 RStudio has 4 main panes, as numbered above: 1) console; 2) source; 3) Environment; and 4) Files, Plots, Packages, etc. 
 
 1. The "console" is where you can run commands just as though you were working in an R environment at a command line, and it is also where results will print out. 
[…]
