+++ R/more_indexing.md
@@ -9,141 +9,174 @@ permalink: /R/more_indexing
 
 {% include _side_tab_R.html %}
 
-Okay, so some of the awesome utility of R is how easy it makes it to subset things down to whatever you want. The fundamentals of this were covered [here](/R/basics#the-wonderful-world-of-indexing) (and if you're completely new to R be sure to run through the [basics](/R/basics) before this), but now we're going to go a bit deeper into the weeds of how exactly subsetting by [conditional statements](/R/basics#subsetting-by-conditional-statements) works. Having a better grasp of what R's doing under the hood will greatly improve your abilities to write and make use of more complicated subsetting expressions.
+Okay, so some of the awesome utility of R is how easy it makes it to subset things down to whatever we want. The fundamentals of this were covered [here](/R/basics#the-wonderful-world-of-indexing){:target="_blank"} (and if you're completely new to R be sure to run through the [basics](/R/basics){:target="_blank"} before this), but now we're going to go a bit deeper into the weeds of how exactly subsetting by [conditional statements](/R/basics#subsetting-by-conditional-statements){:target="_blank"} works. Having a better grasp of what R's doing under the hood will greatly improve our abilities to write and make use of more complicated subsetting expressions.
 <br>
 <br>
 
 ---
 ---
 <br>
+
+# Accessing our R environment
+Before we get started, we need an R environment to work in. You can work on your own computer if you'd like, or you can work in a [Binder](https://mybinder.org/){:target="_blank"} (a self-contained web-accessible environment) that's been created for this page and the [Intro to R page](/R/basics){:target="_blank"}. Both avenues are discussed [here](/R/basics#accessing-our-r-environment){:target="_blank"} if you'd like more info. If you'd like to work in the Binder here, just click this badge to open it: [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/AstrobioMike/binder-R-basics/master?urlpath=rstudio){:target="_blank"} 🙂
 
 # Subsetting by position
 
-Subsetting by index number (the actual sequential position of a value) is pretty straightforward as we covered [here](/R/basics#subsetting-by-position). Let's make a simple vector to work with and then look at that again for a second: 
+Subsetting by index number (the actual sequential position of a value) is pretty straightforward as we covered [here](/R/basics#subsetting-by-position){:target="_blank"}. Let's make a simple vector to work with and then look at that again for a second: 
 
 ```R
 y <- c(5, 6, 7)
 y
 ```
 
-<center><img src="{{ site.url }}/images/R_vector.png"></center> 
 <br>
-And now let's subset it by index values:
+<center><img src="../images/R_vector_ind.png" width="90%"></center> 
+<br>
+
+And now let's subset it by index values (remember these are just the numerical positions, the order, of the values stored in the variable):
 
 ```R
-y # returns the whole vector
-y[1] # returns the first item
-y[2] # returns the second item
-y[3] # returns the third item
-y[c(1,3)] # returns items 1 and 3
+y
+y[1]
+y[2]
+y[3]
+y[c(1,3)]
 ```
 
-<center><img src="{{ site.url }}/images/R_vec_index.png"></center> 
+<br>
+<center><img src="../images/R_vec_ind_2.png" width="90%"></center> 
 <br>
 
-Ok. So each value in a vector has an index number which is just the order of things, and we can use that index number to pull out the value that is held at that position. And we saw how this worked for 2-dimensional structures like tables [here](/R/basics#subsetting-tables).  
[…]
