### 2019-06-22 91630b7
modifying about page, adding 5 things i wish i knew
