### 2019-06-22 735b9d8
massive overhaul of bash section; termed unix everywhere instead; broke down intro into 5 parts; old section archived
