+++ unix/index.md
@@ -1,0 +1,44 @@
+---
+layout: main
+title: Unix home
+categories: [unix]
+permalink: /unix/index
+---
+
+{% include _side_tab_unix.html %}
+
+If you're new to the empowering world of Unix and would like to dive in, you should run through the 5 short pages that comprise the [Unix crash course](/unix/unix-intro){:target="_blank"}. Other than that, feel free to check out the other pages under the Unix pull-down menu above and listed below 🙂
+
+<hr style="height:10px; visibility:hidden;" />
+# Some terminology 
+Here are some terms that are often used interchangeably – not because it's important to remember them or any differences (it's not for most of us), but just to have them laid out somewhere. 
+
+| Term     | What it is          |
+|:-------------:|------------------|
+| **`shell`** | what we use to talk to the computer; anything where you are pointing and clicking with a mouse is a **G**raphical **U**ser **I**nterface (**GUI**) shell; something with text only is a **C**ommand **L**ine **I**nterface (**CLI**) shell |  
+| **`command line`** | a text-based environment capable of taking input and providing output |  
+| **`Unix`** | a family of operating systems |  
+| **`bash`** | the most common programming language used at a Unix command-line |  
+
+<hr style="height:10px; visibility:hidden;" />
+
+---
+---
+<br>
+
+# All Unix pages
+
+* [Getting a Unix environment](/unix/getting_unix_env)
+* [Unix crash course](/unix/unix-intro)
+	1. [Getting started](/unix/getting-started)
+	2. [Working with files and directories](/unix/working-with-files-and-dirs)
+	3. [Redirectors and wildcards](/unix/wild-redirectors)
+	4. [Six glorious commands](/unix/six-glorious-commands)
+	5. [Variables and For loops](/unix/for-loops)  
+
+* [Why is this all worth it? (Some everyday examples)](/unix/why)
+* [What the heck is my PATH, and how do I get it to do what I want?](/unix/modifying_your_path)
+* [Help installing tools](/unix/installing_tools)
+<!--* [Some useful one-liners](/bash/one_liners)-->
+* [Downloading from NCBI](/unix/ncbi_eutils)
+* [Other great resources](/unix/other_resources)  
