+++ R/basics.md
@@ -36,7 +36,7 @@ Part of what makes R so valuable and powerful are all of the open-source package
 Before we get started, we need an R environment to work in. You can work on your own computer if you'd like, or you can work in a "Binder" that's been created for this page, see below. 
 
 ## On your computer
-It is possible your computer already has R, if you are unsure, you can check by opening a terminal window ([a unix-like terminal](/bash/getting_bash_env){:target="_blank"}) and typing **`R`**. If this launches R rather than giving an error message, you should be good to go (enter **`q()`** to exit the R environment). If you do not have R, you can download it from here for Mac: [https://cran.r-project.org/bin/macosx/](https://cran.r-project.org/bin/macosx/){:target="_blank"}. And if you have a relatively newer Mac, you may also need to install XQuartz which you can get from here: [https://www.xquartz.org/](https://www.xquartz.org/){:target="_blank"}. 
+It is possible your computer already has R, if you are unsure, you can check by opening a terminal window ([a unix-like terminal](/unix/getting_unix_env){:target="_blank"}) and typing **`R`**. If this launches R rather than giving an error message, you should be good to go (enter **`q()`** to exit the R environment). If you do not have R, you can download it from here for Mac: [https://cran.r-project.org/bin/macosx/](https://cran.r-project.org/bin/macosx/){:target="_blank"}. And if you have a relatively newer Mac, you may also need to install XQuartz which you can get from here: [https://www.xquartz.org/](https://www.xquartz.org/){:target="_blank"}. 
 
 Lastly, I highly, *highly*, **highly** recommend installing the free version of [RStudio](https://www.rstudio.com/){:target="_blank"} if you don't already have it. RStudio is an interface for R that not only makes everything you will do in R easier and more organized, but it's also invaluable for reproducibility of your analyses as it makes it second-nature to generate and save R scripts of everything you're doing, while you're doing it – which is very helpful when you want to look back and see what worked out of the 20 things you just tried 🙂. You can download an RStudio installer from [here](https://www.rstudio.com/products/rstudio/download/#download){:target="_blank"}. See the next section for the typical layout of RStudio.  
 
@@ -45,12 +45,12 @@ Lastly, I highly, *highly*, **highly** recommend installing the free version of 
 
 When that page finishes loading (it may take a couple minutes), you will see a screen like this:
 
-<center><img src="{{ site.url }}/images/binder-R-launch.png"></center>
+<center><img src="../images/binder-R-launch.png"></center>
 <br>
 
 If we then click the icon with the green plus sign at the top-left, and then "R Script", it will open up our text editor area also, and look something like this (minus the labels added here): 
 
-<center><img src="{{ site.url }}/images/binder-R-labeled.png"></center>
+<center><img src="../images/binder-R-labeled.png"></center>
 
 <br>
 
@@ -70,7 +70,7 @@ Here we're going to be doing our work in the "console" area. To start, let's see
 ?getwd
 ```
 
-<center><img src="{{ site.url }}/images/binder-R-getwd-help.png"></center>
+<center><img src="../images/binder-R-getwd-help.png"></center>
 <br>
 And notice the pane at the bottom right now shows our help info for this function.  
 <br>
@@ -78,7 +78,7 @@ And notice the pane at the bottom right now shows our help info for this functio
 ---
 <br>
 # Some practice data
[…]
