+++ unix/working-with-files-and-dirs.md
@@ -1,0 +1,255 @@
+---
+layout: main
+title: 2. Working with files and directories
+categories: [unix, tutorial]
+permalink: /unix/working-with-files-and-dirs
+---
+
+{% include _unix_working_with_files_and_dirs_toc.html %}
+
+{% include _side_tab_unix.html %}
+
+<hr>
+<center>This is part 2 of 5 of an <a href="/unix/unix-intro" target="_blank">introduction to Unix</a>. If you'd like to follow along, but need to pull up the proper working environment, visit <a href="/unix/getting-started#accessing-our-command-line-environment" target="_blank">here</a> and then come back 🙂</center>
+<hr>
+
+<br>
+
+> **Things covered here:**
+> * Working with files
+> * What a plain-text file is
+> * Introduction to a command-line text editor: **`nano`**
+> * Working with directories
+
+<br>
+
+---
+<br>
+
+# Working with plain-text files and directories
+Next we're going to look at some more ways to learn about and manipulate file and directories at the command line.
+
+>To be sure we are still working in the same place, let's run: 
+>```bash
+>cd ~/unix_intro
+>```
+
+<br>
+## Working with files
+We will often want to get a look at a file to see how it's structured. We've already used a very common tool for peeking at files, the **`head`** command. There is also **`tail`**, which prints the last 10 lines of a file:
+
+```bash
+head example.txt
+tail example.txt
+```
+
+This is especially helpful if a file is particularly large, as **`head`** will just print the first ten lines and stop. This means it will be just as instantaneous whether the file is 10KB or 10GB. 
+
+Another standard useful program for viewing the contents of a file is **`less`**. This opens a searchable, read-only program that allows you to scroll through the document: 
+
+```bash
+less example.txt
+```
+
+To exit the **`less`** program, press the "**q**" key. 
+
+The **`wc`** (**w**ord **c**ount) command is useful for counting how many lines, words, and characters there are in a file: 
+
+```bash
+wc example.txt
+```
+
+<challengeBlock>
+<center><b>QUICK PRACTICE!</b></center>
+
+How can we get <i>only</i> the number of lines in a file from the <htmlCode>wc</htmlCode> command?
+<br>
+
+<div class="wrap-collabsible">
+  <input id="q1" class="toggle" type="checkbox">
+  <label for="q1" class="lbl-toggle">Solution</label>
+  <div class="collapsible-content">
+    <div class="content-inner">
+		<pre>wc -l example.txt</pre>
+
+Adding the optional flag <htmlCode>-l</htmlCode> will print just how many lines are in a file. We could find this out by running <htmlCode>wc --help</htmlCode> or by visiting our good friend Google 🙂 
+<br>
+<br>
+<center>Printing out how many lines are in a file like this is super-useful for things like quickly seeing how many rows are in a large table, or how many sequences are in a file.</center>
+		
+    </div>
+  </div>
+</div>
+</challengeBlock>
+<br>
+
+The most common command-line tools like these and many others we'll see are mostly only useful for operating on what are known as **plain-text files** – also referred to as "flat files". 
+
+### BONUS ROUND: What's a plain-text file?
+A general definition of a plain-text file is a text file that doesn't contain any special formatting characters or information, and that can be properly viewed and edited with any standard text editor.  
+
+Common types of plain-text files are those ending with extensions like ".txt", ".tsv" (**t**ab-**s**eparated **v**alues), or ".csv" (**c**omma **s**eparated **v**alues). Some examples of common file types that are *not* plain-text files would be ".docx", ".pdf", or ".xlsx". This is because those file formats contain special types of compression and formatting information that are only interpretable by programs specifically designed to work with them.  
+
+> **A note on file extensions**<br>
[…]
