+++ unix/other_resources.md
@@ -10,7 +10,7 @@ permalink: /unix/other_resources
 
 Google is always a great place to start if we have a specific problem we are hitting. It's most often going to be something others have already ran into and discussed 🙂
 
-Super helpful sites with excellent forums include [Biostars](https://www.biostars.org/){:target="_blank"} and [StackExchange](https://stackoverflow.com/){:target="_blank"}.
+Super-helpful sites with excellent forums include [Biostars](https://www.biostars.org/){:target="_blank"} and [StackExchange](https://stackoverflow.com/){:target="_blank"}.
 
 [Command-line bootcamp](http://rik.smith-unna.com/command_line_bootcamp/){:target="_blank"} setup by [Rik Smith-Unna](https://twitter.com/blahah404){:target="_blank"} provides an awesome self-contained bash environment alongside tutorials.
 
