### 2019-06-23 9c81d87
Update paper.md
updating to reformatted unix introduction lessons (implemented during the review process as JOSE)
