### 2019-06-23 #53 @AstrobioMike
Regarding the newly generated article proof:

Thanks to input from @JasonJWilliamsNY and in tandem with converting some materials into more of a teaching format for some upcoming workshops (including smaller blocks/sections and interim, collapsible-box-style challenge questions), there's been a pretty large overhaul (in structure, not content) of the Unix and R sections.

So I altered a few sentences in the main text of the paper to reflect this :)

Thanks all!
