### 2019-06-23 #53 @AstrobioMike
Oh, and will definitely need to update the version/zenodo-stored release when we're done here - this is mostly a note to myself :)
