### 2019-06-28 829b9d1
fixing amplicon overview image
