### 2019-06-29 #53 @luizirber
Two issues:

## General checks

- **License**: Is the license for the content GPLv3? CC is recommended for content.

## Documentation

- **Community guidelines**: The issue templates are very good, but maybe adding
  CONTRIBUTING.md also helps for setting expectations when someone opens a pull
  request (there are also PR templates, but that might be overkill =] )

Other than that, I submitted a [PR](https://github.com/AstrobioMike/AstrobioMike.github.io/pull/9) (and it was already merged!) for a typo in the paper, and @JasonJWilliamsNY did a stellar job reporting issues on the original repo, so I think this is good to go.
