+++ stamps2019/getting-started.md
@@ -1,0 +1,295 @@
+---
+layout: main
+title: 1. Getting started
+categories: [unix, tutorial]
+permalink: /stamps2019/getting-started
+---
+
+{% include _unix_getting_started_toc.html %}
+
+<hr>
+<center>This is part 1 of 5 of an <a href="/stamps2019/" target="_blank">introduction to Unix</a>.</center>
+<hr>
+
+> **Things covered here:**
+> * Accessing our Unix-like environment
+> * Some general rules
+> * Running commands and general syntax
+> * File-system structure and how to navigate
+
+---
+<br>
+# How to access the shell for now
+
+To start, we are going to link into our cloud computers stored in this [spreadsheet here](https://hackmd.io/@astrobiomike/stamps2019){:target="_blank"}. Find your name and click the "Jupyter" link associated with it. 
+
+That will open a screen where you will need to enter a password (which we'll tell you the password in the room!) Then a screen like this will open (minus the blue arrows):
+
+<center><img src="../images/shell_tut_jupyter_initial_launch.png" width="90%"></center>
+<br>
+
+Now click the **files tab** at the top-left (that the smaller blue arrow points to above) and then the "**Terminal**" icon at the bottom, and we'll be in our appropriate command-line environment:
+
+<center><img src="../images/shell_stamps_tut_jupyter_terminal.png" width="90%"></center>
+<br>
+
+This is our "command line", and that one line showing with the cursor ready is our "prompt" – where we will be typing all of our commands 🙂 
+
+---
+
+<blockquote><center><b>Important note!</b></center>
+Keep in mind, none of this is about memorization. It may seem counterintuitive at first, but the little details aren't important. What matters is starting to build a mental framework of the foundational rules and concepts. That equips us to figure out the things we need to do, when we need to do them!  
+</blockquote>
+
+---
+<br>
+# A few foundational rules
+
+* **Spaces are special!** The command line uses spaces to know how to properly break things apart. This is why it's not ideal to have filenames that contain spaces, but rather it's better to use dashes ( **`-`** ) or underscores ( **`_`** ) – e.g., "draft_v3.txt" is preferred over "draft v3.txt".  
+
+* The general syntax working at the command line goes like this: **`command argument`**.  
+
+* Arguments can be **optional** or **required** based on the command being used.  
+
+Before we start, copy and paste this one command and hit `enter`, so we are all starting in the same place:
+
+```bash
+cd ~/shell_intro/
+```
+
+---
+
+<blockquote><center><b>COPY AND PASTE STICKY CHECK!</b></center>
+It's okay to copy and paste things here anywhere you'd like. There can of course be some value to typing everything out, but right now is more just about the concepts. How exactly to copy and paste can be different on different machines and keyboard layouts, and sometimes just different because of the terminal we are working in. Let's take a minute now to make sure we can all copy and paste from the material into our terminals. 
+<br><br>
+There may be keyboard shortcuts that work, and/or holding shift and "right-clicking" may bring up the menu too if other ways are not working. 
+</blockquote>
+
+---
+<br>
+# Running commands
+
+**`date`** is a command that prints out the date and time. This particular command does not require any arguments:
+
+```bash
+date
+```
+
+When we run **`date`** with no arguments, it uses some default settings, like assuming we want to know the time in our local time zone. But we can provide optional arguments to **`date`**. Optional arguments most often require putting a dash in front of them in order for the program to interpret them properly. Here we are adding the **`-u`** argument to tell it to report UTC time instead of the local time: 
+
+```bash
+date -u
+```
+
+Note that if we try to run it without the dash, we get an error:
+
+```bash
+date u
+```
+
[…]
