+++ stamps2019/unix-crash-course.md
@@ -1,0 +1,79 @@
+---
+layout: main
+title: Unix crash course
+categories: [unix]
+permalink: /stamps2019/
+---
+
+**Familiarity with the Unix command-line is very likely the most foundational skillset we can develop for bioinformatics** (and for many computational arenas beyond bioinformatics). This is a set of 5 introductory tutorials to help us get from being completely new to Unix up to being great friends with it 🙂 
+
+<hr style="height:15px; visibility:hidden;" />
+
+---
+<br>
+
+# Some terminology 
+Here are some terms that are often used interchangeably – not because it's important to remember them or any differences (it's not for most of us), but just to have them laid out somewhere. 
+
+| Term     | What it is          |
+|:-------------:|------------------|
+| **`command line`** | a text-based environment capable of taking input and providing output |  
+| **`shell`** | what we use to talk to the computer; anything where we are pointing and clicking with a mouse is a **G**raphical **U**ser **I**nterface (**GUI**) shell; something with text only is a **C**ommand **L**ine **I**nterface (**CLI**) shell |  
+| **`Unix`** | a family of operating systems |  
+| **`bash`** | the most common programming language used at a Unix command-line |  
+
+<hr style="height:15px; visibility:hidden;" />
+
+---
+<br>
+
+# Why learn the Unix shell?
+*  enables the use of non-GUI (Graphical User Interface) tools
+*  reproducibility
+*  quickly perform operations on large files (without reading into memory)
+*  automation of repetitive tasks (need to rename 1,000 files?)
+*  enables use of higher-powered computers elsewhere (server/cloud)  
+*  it's the foundation for most of bioinformatics
+
+<hr style="height:15px; visibility:hidden;" />
+
+---
+<br>
+
+# Before we start
+<hr style="height:15px; visibility:hidden;" />
+
+<center><b>My journey into bioinformatics</b></center>
+<center><img src="../images/mike_philosophy2.png" title="Don't worry about every little detail!" width="90%"></center>
+<br>
+
+>**NOTE:** Maybe the most imporant thing to keep in mind while going through these pages is that this is all about <i>exposure</i>, not memorization or mastering anything. Don't worry about the details! Starting to build a foundation of fundamentals will allow us to figure out more details for specific things when we need to 🙂  
+
+
+<hr style="height:15px; visibility:hidden;" />
+
+---
+<br>
+
+
+# Let's get to it!
+1. [Getting started](/stamps2019/getting-started)
+2. [Working with files and directories](/stamps2019/working-with-files-and-dirs)
+3. [Redirectors and wildcards](/stamps2019/wild-redirectors)
+
+---
+
+<hr style="height:10px; visibility:hidden;" />
+<center><b>Return to these later tonight if you'd like 🙂</b></center>
+
+<ol start="4">
+  <li><a href="/stamps2019/six-glorious-commands">Six glorious commands</a></li>
+  <li><a href="/stamps2019/for-loops">Variables and for loops</a></li>
+</ol>
+
+<hr style="height:10px; visibility:hidden;" />
+
+---
+---
+
+<h5><a href="/stamps2019/getting-started" style="float: right"><b>Next:</b> 1. Getting started</a></h5>
