+++ misc/index.md
@@ -20,18 +20,25 @@ This is a [command-line example of processing an amplicon dataset](/amplicon/wor
 
 ---
 <br>
-# STAMPS 2018
+# STAMPS
 The [STAMPS](https://stamps.mbl.edu/index.php/Main_Page){:target="_blank"} course (Strategies and Techniques for Analyzing Microbial Population Structures) at the Marine Biological Laboratory in Woods Hole, MA, USA.
 
+## STAMPS 2018
 ### Introduction to Unix
 The basics to working at the command line along with some definitions and philosophy of starting out in bioinformatics can be found [here](https://astrobiomike.github.io/stamps2018/unix_intro){:target="_blank"}.
 
 ### Amplicon and Metagenomics overviews
 [This page](/misc/amplicon_and_metagen){:target="_blank"} presents a broad-level overview of amplicon sequencing and metagenomics as approaches for microbial ecology, discussing some things each is useful and _not_ useful for, and then outlining some general workflows for each.
+
+## STAMPS 2019
+### Introduction to Unix
+The basics to working at the command line along with some definitions and philosophy of starting out in bioinformatics can be found [here](/stamps2019/unix_intro){:target="_blank"}.
 <br>
 
 ---
 <br>
+
+
 # DIBSI 2018
 [DIBSI](http://ivory.idyll.org/dibsi/){:target="_blank"} (Data Intensive Biology Summer Institute) has been a bioinformatics workshop run by Titus Brown for almost a decade. It now happens at UC Davis, USA each year. 
 

[…]
