+++ stamps2019/unix_intro.md
@@ -1,722 +1,0 @@
----
-layout: main
-title: Introduction to Unix
-categories: [bash, tutorial]
-permalink: /stamps2019/unix_intro
----
-
-{% include _unix_toc.html %}
-
-> **Things covered here**  
-*  Why familiarity with the command line is valuable
-*  Running commands and general syntax
-*  File-system structure and how to navigate
-*  Viewing, creating, and manipulating "plain-text" documents
-*  Intro to pipes and redirectors
-*  Intro to wildcards
-
-<br>
-**Most importantly, none of this is about memorization. It may seem counterintuitive, but the minute details aren't important. What matters is starting to build a mental framework of the foundational rules and concepts. That equips us to figure out the things we need to do, when we need to do them!** 
-
-<br>
-<center><img src="../images/mike_philosophy2.png" title="Don't worry about every little detail!" width="90%"></center>
-
-<br>
-
----
-<br>
-
-# Some terminology
-
-| Term     | What it is          |
-|:-------------:|------------------|
-| **`Unix`** | a family of operating systems |
-| **`command line`** | a text-based environment capable of taking input and providing output |
-| **`shell`** | our ambassador to the operating system; this translates between us and the computer |
-| **`bash`** | the most common programming language used at a Unix command-line |  
-
-<br>
-
-# Why learn the command line?
-
-*  it's the foundation for most of bioinformatics
-*  enables use of non-GUI (Graphical User Interface) tools
-*  quickly perform operations on large files
-*  reproducibility
-*  automation of repetitive tasks
-	*  need to rename 1,000 files?
-*  enables use of higher-powered computers elsewhere (server/cloud)  
-
-<br>
-
----
-<br>
-
-# Connect to the MBL server
-We are going to be doing most of our work during the course on the MBL server. So before we get started, we are going to login there. So open your computer's command-line environment and let's login following these instructions.
-
-On your badge, you have a username and password. That will be your account for the course. 
-
-To connect, we'll use **ssh**, a command that let's us establish a secure connnection. 
-
-> **If you are on the MBL-GUEST or eduroam WiFi network, start with Step 1, and then go to Step 2. If you are on the MBL-REGISTERED WiFi network, skip Step 1 and start at Step 2.**
-
-**Step 1**  
-At the command line type the following, but replace "\<USERNAME\>" with your username:
-
-```
-ssh <USERNAME>@class.mbl.edu
-```
-
-Don't include the brackets. For example, if your username were "mlee", like mine is, the command would look like `ssh mlee@class.mbl.edu`. 
-
-The first time you log in you will get a message about security. Type "yes", and press `enter`. 
-
-Then it will ask you for your password. Type the password listed on your badge and press `enter`; the cursor will not respond, but it is still being entered as you type.
-
-> **Pro tip**  
-> If you have a crazy password, you can type it somewhere you can read it. Then copy and paste it.
-
-**Step 2**  
-So far, we just logged into what is called a 'gateway'. We won't be doing our work here though, we're going to take one more step and log into specific servers so we don't overload the computing resources. 
-
-On your badge you also have something that says a class and a number, e.g. "class-09". That specifies which server you should be working on. So now we want to `ssh` into that specific server.  so enter the following, again change the "\<USERNAME\>" to your username, and change the "\<??>\" to your specific class number:
-
-```
-ssh <USERNAME>@class-<??>
-``` 
-
-Again, don't include the brackets. For example, if my class number is "class-09", the command would look like: `ssh mlee@class-09`.
-
-Then if it asks for you password again, enter it the same way as above and press `enter`.
-
-Now you're logged into the server that you'll use for much of the course.
-
-> **NOTE**  
[…]
