+++ stamps2019/six-glorious-commands.md
@@ -8,7 +8,7 @@ permalink: /stamps2019/six-glorious-commands
 {% include _unix_six_commands_toc.html %}
 
 <hr>
-<center>This is part 4 of 5 of an <a href="/stamps2019/unix-intro" target="_blank">introduction to Unix</a>. If you'd like to follow along, but need to pull up the proper working environment, visit <a href="/stamps2019/getting-started#accessing-our-command-line-environment" target="_blank">here</a> and then come back 🙂</center>
+<center>This is part 4 of 5 of an <a href="/stamps2019/unix-crash-course" target="_blank">introduction to Unix</a>. If you'd like to follow along, but need to pull up the proper working environment, visit <a href="/stamps2019/getting-started#accessing-our-command-line-environment" target="_blank">here</a> and then come back 🙂</center>
 <hr>
 
 <hr style="height:15px; visibility:hidden;" />
