+++ stamps2019/working-with-files-and-dirs.md
@@ -29,7 +29,7 @@ permalink: /stamps2019/working-with-files-and-dirs
 
 >To be sure we are still working in the same place, let's run: 
 >```bash
->cd ~/unix_intro
+>cd ~/shell_intro
 >```
 
 <hr style="height:5px; visibility:hidden;" />
