+++ stamps2019/for-loops.md
@@ -28,7 +28,7 @@ Loops are extremely powerful in all programming languages. They are what let us 
 
 >To be sure we are still working in the same place, let's run: 
 >```bash
->cd ~/unix_intro
+>cd ~/shell_intro
 >```
 
 <hr style="height:10px; visibility:hidden;" />
