### 2019-07-25 #53 @AstrobioMike
Thanks for the call on the license, @luizirber! I think i just chose GPL because github doesn't nicely let you select a CC license and i didn't look any further at the time, and didn't really understand the differences much. Now it's a CC-BY-4 :)

And while there aren't PR templates, there are Issue templates now thanks to a suggestion by @JasonJWilliamsNY :) https://github.com/AstrobioMike/AstrobioMike.github.io/issues/new/choose

And I've added a Contribution.md https://github.com/AstrobioMike/AstrobioMike.github.io/blob/master/Contributing.md

Thanks all!
