### 2019-07-26 19b0f3c
update amplicon metagen page
