+++ misc/amplicon_and_metagen.md
@@ -6,75 +6,80 @@ permalink: /misc/amplicon_and_metagen
 
 {% include _amp_and_meta_toc.html %}
 
-This page presents a broad-level overview of amplicon sequencing and metagenomics as approaches for microbial ecology and outlines some generic workflows for each listing some relevant tools. Both of these methods are just two more tools of investigation and should be thought of as steps in the process of science rather than end-points (like all tools of science). They are most often applied for exploration and hypothesis generation. 
-
-
-> "Why are you doing 16S sequencing? That doesn't tell you anything about function."  
-> 
-> "Why are you measuring nitrogen-fixation rates? That doesn't tell you anything about the biochemistry doing it."  
-> 
-> **Don't assess the utility of a tool based on something it's not supposed to do anyway.**
-
-<br>
-
-___
-<br>
+This page presents a broad-level overview of amplicon sequencing and metagenomics as applied to microbial ecology. Both of these methods are most often applied for exploration and hypothesis generation and should be thought of as steps in the process of science rather than end-points – like all tools of science 🙂 
 
 **Amplicon sequencing**  
-Amplicon sequencing of marker-genes (e.g. 16S, 18S, ITS) involves using specific primers that target a specific gene or gene fragment. It is one of the first tools in the microbial ecologist's toolkit. It is most often a broad-level survey of community composition used to generate hypotheses based on differences between samples (***differences based on recovered gene-copy numbers, not abundance of organisms***).
+
+Amplicon sequencing of marker-genes (e.g. 16S, 18S, ITS) involves using specific primers that target a specific gene or gene fragment. It is one of the first tools in the microbial ecologist's toolkit. It is most often used as a broad-level survey of community composition used to generate hypotheses based on differences between recovered gene-copy numbers between samples.
 
 **Metagenomics**  
-Shotgun metagenomic sequencing provides a way to access *all* the DNA of a mixed community. It uses random rather than targeted primers and therefore suffers much less from pcr bias. (It still suffers from other things such as cell-lysis rates which can be dependent on organism cell walls and the extraction method used.)
 
-<br>
+Shotgun metagenomic sequencing aims to amplify all the accessible DNA of a mixed community. It uses random primers and therefore suffers much less from pcr bias (discussed below). Metagenomics enables profiling of taxonomy and functional potential. Recently, the recovery of representative genomes from metagenomes has become a very powerful approach in microbial ecology, drastically expanding the known Tree of Life by granting us genomic access to as-yet unculturable microbial populations (e.g. [Hug et al. 2016](https://www.nature.com/articles/nmicrobiol201648); [Parks et al. 2017](https://www.nature.com/articles/s41564-017-0012-7)). 
 
-___
-<br>
+Here we'll discuss some of the things each is useful and not useful for, and then look at some general workflows for each. 
 
-# Some capabilities of each
-<br>
 
-## Amplicon
+## Amplicon sequencing utility
 * **Useful for:**
-
-    * one *metric* of community composition
-        * can say something about relative abundance of gene copies recovered (recovered gene copies ≠ counts of organisms)
-        * gives a snapshot of, e.g., "16S gene-fragment copies recovered"
-    * can track changes in community structure (as interpreted by recovered gene copy numbers) in response to a treatment and/or across environmental gradients/time, etc.
-    * can provide strong support for further investigation, particularly single-nucleotide resolution methods
-        * e.g. *Trichodesmium–Alteromonas* story ([starting paper here](https://www.nature.com/articles/ismej201749){:target="_blank"})  
-<br>
+    * one metric of community composition
+        * can say something about relative abundance of gene copies recovered
[…]
