### 2019-07-30 5d65d9a
research update
