### 2019-08-21 7dbb865
incomplete ssh page for stamps 2019
