+++ stamps2019/ssh.md
@@ -1,0 +1,388 @@
+---
+layout: main
+title: Connecting to a remote computer with ssh
+categories: [unix, tutorial]
+permalink: /stamps2019/ssh
+---
+
+<hr style="height:10px; visibility:hidden;" />
+
+> **Things covered here:**
+> * Here we are going to look at how to connect to a remote computer from our command line using a program called `ssh` (**s**ecure **sh**ell).
+
+<hr style="height:10px; visibility:hidden;" />
+
+---
+<hr style="height:25px; visibility:hidden;" />
+
+To do this as covered below, you need to have a working [Unix-like environment](/unix/getting_unix_env).
+
+<hr style="height:10px; visibility:hidden;" />
+
+---
+<hr style="height:25px; visibility:hidden;" />
+
+We'll mostly be working with a file here called "gene_annotations.tsv", which is a tab-delimited table holding genes, their annotations, and their amino acid sequences. To help orient us, here is a peek at it in Excel:
+
+<center><img src="../images/unix_example_gene_annots.png" width="70%"></center>
+<br>
+
+We can see the file has 4 columns: "gene_ID", "genome", "KO_ID", and "KO_annotation" (KO is for **K**egg **O**rthology, a popular functional database). Let's change into our working directory for this page and explore it a little at the command line with some of the things we've seen already.
+
+```bash
+cd ~/shell_intro/six_commands/
+head gene_annotations.tsv
+```
+
+We can also just pull out the header if we'd like by setting the number of lines **`head`** prints out to 1:
+
+```bash
+head -n 1 gene_annotations.tsv
+```
+
+And let's count how many rows there are:
+
+```bash
+wc -l gene_annotations.tsv
+```
+
+Now that we know something about the file we're working with, let's get to some new commands!
+
+<hr style="height:10px; visibility:hidden;" />
+# cut  
+**`cut`** is a command that's great for manipulating columns. The required arguments for **`cut`** are which columns we want, and from which file. Here's how we can use **`cut`** to pull out just the gene_ID column (column 1):
+
+```bash
+cut -f 1 gene_annotations.tsv
+```
+
+That is printing out all of the lines to the screen though, let's pipe **`|`** it into **`head`** for now to keep things manageable while we're working on it (remember we can bring up a previous command by pressing up):
+
+```bash
+cut -f 1 gene_annotations.tsv | head
+```
+
+Here we are specifying which column we want with the **`-f`** parameter (for **f**ield). We can specify multiple individual columns if we separate them with a comma:
+
+```bash
+cut -f 1,3 gene_annotations.tsv | head
+```
+
+And a range of columns if we join them with a dash:
+
+```bash
+cut -f 1-3 gene_annotations.tsv | head
+```
+
+Just like in Excel, **`cut`** thinks about where columns start and stop based on a delimiter, like a comma or a tab. The default setting is a tab, so we didn't need to change it for that file. But if we try using it on a **c**omma **s**eparated **v**alues file (a csv), things don't work well:
+
+```bash
+cut -f 1-3 example_gene_annotations.csv | head
+```
+
+Unless we tell **`cut`** the delimiter is a comma, which we can do with the **`-d`** parameter:
+
+```bash
+cut -d , -f 1-3 example_gene_annotations.csv | head
+```
+
+<hr style="height:10px; visibility:hidden;" />
+<challengeBlock>
+<center><b>QUICK PRACTICE!</b></center>
+
[…]
