### 2019-08-31 #53 @labarba
@JasonJWilliamsNY — You have a couple of unchecked items in your checklist. Are you satisfied with these items?
