### 2019-08-31 #53 @labarba
@AstrobioMike — One important thing I miss from the paper is some mention of other similar efforts, tutorials or content. In other words, a light overview of the state of the art, with some citations. Perhaps you could even mention how your contribution is different to some others. Surely the Software Carpentry and Data Carpentry collections have some bioinformatics tutorials? We even have another JOSE paper on a related topic: https://doi.org/10.21105/jose.00027
