### 2019-08-31 04f3e88
java script update - trying to fix dropdown menus on mobile
