### 2019-09-01 #53 @JasonJWilliamsNY
> @JasonJWilliamsNY — You have a couple of unchecked items in your checklist. Are you satisfied with these items?

Yes - sorry these were somehow missed!
