### 2019-09-02 f07f307
overhauling EDirect page
