### 2019-09-02 f08a4ca
updating tool page and all tutorials index
