+++ unix/installing_tools.md
@@ -18,7 +18,7 @@ permalink: /unix/installing_tools
 >
 >To get conda up and running (which is very quick), you can follow the instructions to install miniconda (a light-weight version) for your appropriate system starting from [here](https://conda.io/en/latest/miniconda.html){:target="_blank"}.  
 >
->The rest of this page discusses how to do some things in a non-conda manner.
+>Some of the install below I've updated to include the conda installation, but others I haven't gotten to yet. Be sure to check if there is a conda installation available first if you can use that (there is for everything here I think).
 
 <hr style="height:10px; visibility:hidden;" />
 
@@ -44,7 +44,7 @@ You may have heard or read an expression before referring to whether there is a 
 ---
 <br>
 # First we need a happy bin
-Often a big part of getting things to work properly is having them in a location on the computer that you can access no matter [where you are](/unix/getting-started#the-unix-file-system-structure){:target="_blank"}. As we covered [here](/unix/modifying_your_path){:target="_blank"}, a list of directories that are scanned for programs automatically by your computer is stored in the special variable called "PATH". For the sake of simplicity, many tools we use in tutorials on this site are going to be put in a directory we made and added to our PATH in the [modifying your PATH walkthrough](/unix/modifying_your_path){:target="_blank"} called `~/happy_bin`. You can check to see if this is in your PATH already like this:
+Often a big part of getting things to work properly is having them in a location on the computer that you can access no matter [where you are](/unix/getting-started#the-unix-file-system-structure){:target="_blank"}. As we covered [here](/unix/modifying_your_path){:target="_blank"}, a list of directories that are scanned for programs automatically by your computer is stored in the special variable called "PATH". For the sake of simplicity, many tools we use in tutorials on this site are going to be put in a directory called `~/happy_bin`. You can check to see if this is in your PATH already like this:
 
 ```
 echo $PATH | tr ":" "\n" | grep "happy_bin"
@@ -59,7 +59,7 @@ You're good to go. If nothing was returned, then you can create this directory a
 
 ```
 mkdir ~/happy_bin
-echo 'export PATH="$PATH:/Users/Mike_Lee/happy_bin"' >> ~/.bash_profile
+echo 'export PATH="$PATH:~/happy_bin"' >> ~/.bash_profile
 ```
 
 And if you're unsure of what's going on here, be sure to visit the [modifying your PATH page](/unix/modifying_your_path){:target="_blank"}.  
@@ -70,11 +70,31 @@ And if you're unsure of what's going on here, be sure to visit the [modifying yo
 <br>
 # Installing tools used on this site
 As explained [above](/unix/installing_tools#first-we-need-a-happy-bin){:target="_blank"}, most things we install here will be put in a directory in our home location called `~/happy_bin`, so pay attention to modify any code here accordingly if you want to put things somewhere else. If you're unfamiliar with what's going on in any of these code blocks below, and want to be more familiar, run through the [Unix crash course](/unix/unix-intro){:target="_blank"} and [modifying your PATH](/unix/modifying_your_path){:target="_blank"} pages sometime 🙂  
+
 <hr style="height:10px; visibility:hidden;" />
 
-## My BioInformatics Tools (bit)
[…]
