+++ unix/ncbi_eutils.md
@@ -18,6 +18,8 @@ There is a lot of info on this at [the main NCBI page for this here](https://www
 
 This page holds some of the ways I've used EDirect, both to serve as a handy archive for myself, and to hopefully help others 🙂 It won't be as comprehensive as most other things on this site, as it's extremely expansive and it's still nowhere near intuitive for me 🤷‍♂️ But these examples may do what is needed, and if not they at least might provide good starting points for building the code that will do what is needed. 
 
+>**NOTE:** This stuff can look messy, mostly because it is. This page assumes you already have some familiarity with working at the command line. If you don't yet, then definitely consider running through the [Unix crash course](/Unix/unix-intro){:target="_blank"} first 🙂
+
 <hr style="height:15px; visibility:hidden;" />
 
 ---
@@ -38,7 +40,7 @@ conda install -y -c conda-forge -c bioconda -c defaults entrez-direct
 <hr style="height:15px; visibility:hidden;" />
 ## Accessing genome assemblies and info
 
-* **All *Alteromonas* assembly accessions** (for instance to input into [GToTree](https://github.com/AstrobioMike/GToTree/wiki/what-is-gtotree%3F){:target="_blank"} like the [example here](https://github.com/AstrobioMike/GToTree/wiki/example-usage#alteromonas-example){:target="_blank"}!)
+**Getting all *Alteromonas* assembly accessions** (e.g. to input into [GToTree](https://github.com/AstrobioMike/GToTree/wiki/what-is-gtotree%3F){:target="_blank"} like the [example here](https://github.com/AstrobioMike/GToTree/wiki/example-usage#alteromonas-example){:target="_blank"}!)
 
 ```bash
 esearch -db assembly -query '"Alteromonas"[Organism] AND latest[filter] AND \
@@ -51,7 +53,7 @@ esearch -db assembly -query '"Alteromonas"[Organism] AND latest[filter] AND \
 
 <hr style="height:10px; visibility:hidden;" />
 
-* **All RefSeq Bacteria assembly accessions, taxids, assembly status, number of contigs, L50, N50, and total assembly length** (took ~5 minutes to get 166,566 records as accessed on 1-Sep-2019)
+**Getting all RefSeq Bacteria assembly accessions, taxids, assembly status, number of contigs, L50, N50, and total assembly length** (took ~5 minutes to get 166,566 records as accessed on 1-Sep-2019)
 
 ```bash
 esearch -db assembly -query '"Bacteria"[Organism] AND "latest refseq"[filter] AND \
@@ -66,9 +68,18 @@ esearch -db assembly -query '"Bacteria"[Organism] AND "latest refseq"[filter] AN
 
 <hr style="height:15px; visibility:hidden;" />
 
-* **Downloading genomes by accession**  
+**Downloading genomes by accession**  
 
-I'd typically do this part with one of the tools listed above in the intro, after getting the accessions like the examples above, but here's an example pulling the sequence data for a RefSeq accession:
+I'd typically do this part using `bit-dl-ncbi-assemblies` from my [Bioinf Tools](https://github.com/AstrobioMike/bioinf_tools#bioinformatics-tools-bit){:target="_blank"} after getting the accessions like in the examples above like so:
+
+```bash
+echo -e "GCF_006538345.1\nGCF_006538325.1\nGCF_006538305.1\nGCF_006517115.1" \
+     > assembly-accs.txt
+
+bit-dl-ncbi-assemblies -w assembly-accs.txt -f fasta -j 4
+```
+
+But here's an example using EDirect to pull the sequence data for a RefSeq accession:
 
 ```bash
[…]
