### 2019-09-03 170dd20
adding links to all pages to each index page because i can't get the damn dropdown menus to work on mobile :(
