+++ genomics/index.md
@@ -12,3 +12,13 @@ categories: [genomics]
 <p align="right"><a href="https://ndownloader.figshare.com/files/15628103">PDF download</a></p>
 
 Genomics is a very broad subject. In my small bubble, my time spent in the genomics realm involves mostly working with bacteria and archaea doing things like assembly, functional annotation, assessing distributions across metagenomes, comparative genomics, and trying to infer evolutionary relationships through phylogenomics. But as usual there are a ton of questions to pursue and lots of ways to pursue them. Here I'm going to provide some examples of some of the things I do 🙂
+
+<hr style="height:10px; visibility:hidden;" />
+
+---
+<br>
+# Genomics pages
+
+* [Where to start: quality filtering, error correction, and depth normalization](/genomics/where_to_start)  
+* [*de novo* genome assembly](/genomics/de_novo_assembly) 
+* [Recovering **M**etagenome-**A**ssembled **G**enomes (MAGs) from metagenomes](/genomics/metagen_anvio)
