+++ R/index.md
@@ -11,11 +11,20 @@ permalink: /R/
 [R](https://www.r-project.org/) is generally a big, glorious calculator. More specifically, it's a programming language and working environment for statistical analyses and figure generation, and it's pretty sweet. What really makes R powerful is that it is completely open source and it has a tremendous number of incredible people that contribute "packages" to it for all to use. Packages are bundles of code that perform specific tasks, and they are kinda like Apps in the sense that for most of the things you'll want to do, "there's a package for that". Often you will end up searching for a package that you know exists already either from hearing about it from someone, or seeing it used in a paper, but you can also search the [Comprehensive R Archive Network (CRAN) for packages](https://cran.r-project.org/web/packages/) directly. R serves as the foundation upon which you can utilize this large swath of tools that people all over the world have developed and contributed, and it also is invaluable for parsing tables and creating figures. It's definitely worth the initial time investment it may take to get comfortable with it.  
 
 There is extensive documentation on R at the CRAN [Introduction to R site](https://cran.r-project.org/doc/manuals/r-release/R-intro.html), and it would probably be worthwhile going through it at some point. Consistent with the general approach of the site here, I try to distill things down to just the baseline skills to start in the [R basics]({{ site.url }}/R/basics) page.  
+
 <br>
 
 ---
 <br>
+# R pages
+* [Introduction to R](/R/basics)  
+* [Going deeper with indexing](/R/more_indexing)  
+* [Installing R packages](/R/installing_packages)  
 
+<hr style="height:10px; visibility:hidden;" />
+
+---
+<br>
 <h3>Some terminology</h3>
 
 **Variable**  
