### 2019-09-08 7edaa8d
updating jose paper
