### 2019-09-08 #53 @AstrobioMike
Hi there, @labarba!

I added in some other resources in the second paragraph and worked in highlighting two of the main ways my site is different in the remaining text where appropriate. Thanks for the suggestion :)

I actually reviewed the https://doi.org/10.21105/jose.00027 paper, ha :) I didn't add that in though because that's actually very different in that it doesn't teach biologists how to use bioinformatics from scratch, or really at the user level. Rather it is more teaching people who can already program (particularly in python) how things work conceptually and then how to write bioinformatic algorithms to do that thing - like how multiple sequence alignment works, and then how to write an alignment program, or the same for a neighbor-joining tree, very in-the-weeds stuff – I learned a lot from reviewing it!

Thanks again, all :)
