### 2019-09-09 5c28495
updating full amplicon example page
