+++ amplicon/dada2_workflow_ex.md
@@ -53,7 +53,7 @@ We will be using the command line first, as detailed below, so you can leave it 
 <br>
 
 # Tools used here
-If using the Binder environment, you don't need to worry about installing these tools, but if working on your own system, you will need to take care of some installations as noted while we go. In addition to [DADA2](https://benjjneb.github.io/dada2/index.html){:target="_blank"} as our main processing tool, we'll be using [cutadapt](https://cutadapt.readthedocs.io/en/stable/index.html){:target="_blank"} to remove our primers. DADA2 is available as an [R](https://www.r-project.org/){:target="_blank"} package, with installation instructions provided by the developers [here](https://benjjneb.github.io/dada2/dada-installation.html){:target="_blank"}. If your experience is like mine, it shouldn't give you any trouble if installing on your computer, but you may run into some issues when trying to install on a server where you don't have authorization to do whatever you'd like – a huge thanks to [@phantomBugs](https://twitter.com/phantomBugs){:target="_blank"} for all his help when I was bugging him about that! If you'd like to use [cutadapt](https://cutadapt.readthedocs.io/en/stable/index.html){:target="_blank"}, you can find installation instructions for it from the developers [here](https://cutadapt.readthedocs.io/en/stable/installation.html){:target="_blank"}.
+If using the Binder environment, you don't need to worry about installing these tools, but if working on your own system, you will need to take care of some installations as noted while we go. In addition to [DADA2](https://benjjneb.github.io/dada2/index.html){:target="_blank"} as our main processing tool, we'll be using [cutadapt](https://cutadapt.readthedocs.io/en/stable/index.html){:target="_blank"} to remove our primers. DADA2 is available as an [R](https://www.r-project.org/){:target="_blank"} package, with installation instructions provided by the developers [here](https://benjjneb.github.io/dada2/dada-installation.html){:target="_blank"}. If your experience is like mine, it shouldn't give you any trouble if installing on your computer, but you may run into some issues when trying to install on a server where you don't have authorization to do whatever you'd like – a huge thanks to [@phantomBugs](https://twitter.com/phantomBugs){:target="_blank"} for all his help when I was bugging him about that! If you'd like to use [cutadapt](https://cutadapt.readthedocs.io/en/stable/index.html){:target="_blank"}, you can find installation instructions for it from the developers [here](https://cutadapt.readthedocs.io/en/stable/installation.html){:target="_blank"} (which includes the conda installation method `conda install -c bioconda cutadapt`).
  
 And since DADA2 is an R package, you also of course need to have a working installation of R on your computer. If you'd like/need more info on this, check out the [R basics](/R/basics){:target="_blank"} page before moving forward. 
 <br>
