+++ about/index.md
@@ -6,7 +6,7 @@ permalink: /about/
 
 Hi there 🙂 My name's [Mike](https://twitter.com/AstrobioMike){:target="_blank"}. I'm a bioinformatician who gets to explore microbial ecology and evolution in all kinds of different environments. 
 
-I am the biology-to-bioinformatics type of person. Like so many of us these days, I found myself in a biology lab with lots of bioinformatics work to be done, but little to no guidance or expertise in bioinformatics available. This isn't anyone's fault of course, it's just that **our society as a whole is generating data much faster than it is developing the skills to work with it**. And it takes time for our educational infrastructure to catch up to what is needed. In the meantime, many of us are stuck fighting our way up a pretty steep learning curve trying to develop skills that are rapidly becoming essential for biologists. At some undefined point in gradschool, I made it through the toughest parts of that learning curve, things started getting easier for a change, and I started realizing how glorious and empowering it was on the other side 🙂 
+I am 100% the biology-to-bioinformatics type of person. Like so many of us these days, I found myself in a biology lab with lots of bioinformatics work to be done, but little to no guidance or expertise in bioinformatics available. This isn't anyone's fault of course, it's just that **our society as a whole is generating data much faster than it is developing the skills to work with it**. And it takes time for our educational infrastructure to catch up to what is needed. In the meantime, many of us are stuck fighting our way up a pretty steep learning curve trying to develop skills that are rapidly becoming essential for biologists. At some undefined point in gradschool, I made it through the toughest parts of that learning curve, things started getting easier for a change, and I started realizing how glorious and empowering it was on the other side 🙂 
 
 The purpose of this site is to try to provide the sort of help that would have helped me when I was starting out. Along with the tutorials linked in the dropdown menus above, that also includes some things I wish someone had explicitly told me. So I'll throw them out here 🙂
 
@@ -17,10 +17,9 @@ The purpose of this site is to try to provide the sort of help that would have h
 
 ### 1) Fundamentals and concepts are important, not details
 * Don’t feel like you need to master everything you come across.
-	* This can be very hard for a lot of us!
-* We are not computer scientists, we are biologists using computers.
-	* Meaning, we don’t need to be able to program perfectly in Python in order to use it effectively in our research. Just having a basic foundation is much more attainable alongside the rest of our research and other obligations. And if we need more at some point, we can learn what we need, when we need it. 
-* **Think of it like Excel. Many of us use Excel effectively without knowing how to do everything it can do (which is a lot).**
+	* This can be very hard for a lot of us! But we are not computer scientists, we are biologists using computers.
+	* Meaning, for example, we don’t need to be able to program perfectly in Python in order to use it effectively in our research. Just having a basic foundation is much more attainable alongside the rest of our research and other obligations. And if we need more at some point, we can learn what we need, when we need it. 
+* **Think of it like Excel. Many of us use Excel effectively without knowing how to do everything it can do (which is a lot).** 
 
 <hr style="height:15px; visibility:hidden;" />
 <center><b>High-resolution timelapse of my ongoing journey:</b></center>
@@ -40,24 +39,24 @@ The purpose of this site is to try to provide the sort of help that would have h
 
 * Many of us have little or no bioinformatics guidance in our labs/departments/institutions.
 * It is super-helpful to have a community of people we can connect with that are working on similar things or have already struggled with what we are currently struggling with. 
+	* It's also very valuable just knowing that what we are experiencing **is the norm**!
 * If feasible, in-person workshops are usually filled with people who are excited to help and learn from each other, which makes them a great place to connect.
[…]
