+++ genomics/metagen_anvio.md
@@ -47,7 +47,7 @@ conda install -y bowtie2 anvio diamond
 <br>
 
 # Our practice data
-To work with a smaller dataset here that will let us do things in a reasonable amount of time, we're going to be working with a relatively simple microbial community here that comes from metagenomic sequencing of an enrichment culture of the nitrogen-fixing cyanobacterium *Trichodesmium*. Metagenomics still takes a lot of time, so we're going to start with data already quality trimmed/filtered here, though assessing the quality and trimming/filtering as needed as laid out in [this lesson](https://angus.readthedocs.io/en/2018/quality-and-trimming.html){:target="_blank"} should pretty much always be the first step. To lighten the processing load the majority of *Trichodesmium* (target cultivar) reads have also been removed. Despite this, there are still some steps that would take a bit too long to just wait for, so in those cases there will be examples of how the code would be run, but we'll just pull result files from a subdirectory that comes with the data download so skip some of the more time-consuming steps 🙂
+To work with a smaller dataset here that will let us do things in a reasonable amount of time, we're going to be working with a relatively simple microbial community here that comes from metagenomic sequencing of an enrichment culture of the nitrogen-fixing cyanobacterium *Trichodesmium*. Metagenomics still takes a lot of time, so we're going to start with data already quality trimmed/filtered here, though assessing the quality and trimming/filtering as needed as laid out in [the de novo genome assembly page](https://astrobiomike.github.io/genomics/de_novo_assembly#quality-filtering){:target="_blank"} should pretty much always be the first step. To lighten the processing load the majority of *Trichodesmium* (target cultivar) reads have also been removed. Despite this, there are still some steps that would take a bit too long to just wait for, so in those cases there will be examples of how the code would be run, but we'll just pull result files from a subdirectory that comes with the data download so skip some of the more time-consuming steps 🙂
 
 Downloading the practice data should only take about 3 or 4 minutes (it's ~1.5 GB):
 
