### 2019-09-17 9ef5672
update edirect page
