### 2019-09-25 #53 @labarba
Oops. I forgot the pre-publication steps.

@AstrobioMike : We're going to need you to do the following:

- [ ] Make a tagged release, and list the version tag of the archived version here.
- [ ] Archive the reviewed materials in Zenodo (or similar service)
- [ ] Report the Zenodo DOI of the archived version here.
