### 2019-09-25 #53 @labarba
I found the Zenodo deposit from the DOI caught by `whedon`. Note that the author list needs to be modified, as it now includes one of the reviewers (perhaps who submitted a PR?)
https://doi.org/10.5281/zenodo.3402376
