### 2019-09-25 #53 @labarba
I'm still going to request a few mods, I'm afraid. You've added "Statement of Need" as an italics paragraph header to the first paragraph, which talks about the gap in education, given shifts in science to more data. What I want is a "Statement of Need" for _this learning module that you have developed_. Leave that introductory paragraph there, and find a place for the Statement of Need where you tell us why _your materials_ are useful to others.

I would welcome an effort to remove some clutter in the text: superlatives, adverbs (e.g., fortunately, unfortunately, largely), words like "very" can almost always be removed.

In paragraph 3, you refer to your work as "a website." Please reword this, as JOSE does not publish "websites." We publish fully written out learning modules that can be re-used by other instructors, and by self-learners.
