### 2019-09-26 7e00ba9
paper update
