+++ paper.md
@@ -15,12 +15,12 @@ bibliography: paper.bib
 ---
 
 # Summary
-*Statement of need:* Our civilization's ability to generate data has far surpassed our current, collective capabilities to efficiently access and utilize it. And while the necessary educational infrastructure is trying to adapt and grow as quickly as possible, many graduate students, postdocs, faculty, and other researchers have found ourselves in situations where self-education with little direction is the only option available. This can make it very difficult to get through the initial steep learning curve of beginning to use bioinformatics effectively – particularly when this is expected to happen alongside all of our other research and educational responsibilities.  
+Our civilization's ability to generate data has far surpassed our current, collective capabilities to efficiently access and utilize it. And while the necessary educational infrastructure is trying to adapt and grow as quickly as possible, many graduate students, postdocs, faculty, and other researchers have found ourselves in situations where self-education with little direction is the only option available. This can make it difficult to get through the initial steep learning curve of beginning to use bioinformatics effectively – particularly when this is expected to happen alongside all of our other research and educational responsibilities.  
 
-There are fortunately many resources available online. For example [Codecademy](https://www.codecademy.com/) offers free, online lessons focused on learning several different languages from scratch. And more geared towards the sciences, [The Carpentries](https://carpentries.org/) is an international organization largely powered by volunteers designing and maintaining lessons for in-person workshops that are then openly available. Unfortunately, the plethora of things out there to possibly learn is also very overwhelming when starting out as it can be difficult to know what will be important to us and therefore where to begin. 
+There are many resources available online for learning programming. For example, [Codecademy](https://www.codecademy.com/) offers free, online lessons focused on learning several different languages from scratch. And more geared towards the sciences, [The Carpentries](https://carpentries.org/) is an international organization largely powered by volunteers designing and maintaining lessons for in-person workshops that are then openly available. However, the plethora of things out there to possibly learn is also overwhelming when starting out, as it can be difficult to know what will be important to us and therefore where to begin. And most available resources are lacking an emphasis or any discussion on the philosophy of being a biologist trying to learn bioinformatics – which I've found to be integral to the process.
 
-[Happy Belly Bioinformatics](https://astrobiomike.github.io/) is a website designed and maintained specifically around the goal of helping biologists be able to learn and apply bioinformatics as needed in our research. It has been directly born out of my personal experience of, like many of us these days, being a biologist thrust into the computational world with little or no guidance/advice/resources in my surrounding lab/department/environment. The underlying educational philosophy of the site is that learning a small set of baseline, fundamental skills and concepts that will directly aid in our research is what is important – rather than, for example, sitting down and trying to master a specific programming language without any direct application to our work.
+*Statement of need:* [Happy Belly Bioinformatics](https://astrobiomike.github.io/) is a resource designed and maintained specifically around the goal of helping biologists be able to learn and apply bioinformatics as needed in our research. It has been directly born out of my personal experience of, like many of us these days, being a biologist thrust into the computational world with little or no guidance/advice/resources in my surrounding lab/department/environment – it is the resource I wish I had. The underlying educational philosophy is that learning a small set of baseline, fundamental skills and concepts that will directly aid in our research is what is important – rather than, for example, sitting down and trying to master a specific programming language without any direct application to our work. 
 
[…]
