### 2019-09-26 #53 @AstrobioMike
Thanks, @labarba!

Good point on the statement of need. It's of course all of it (the need stated in the first paragraph is why any of these resources exist), but it definitely makes more sense in this context moved down a bit and with some more wording contrasting this resource vs others.

I'm definitely not a fan of "very"s either in scientific writing. All of the bioinf education stuff is very much in casual communication mode, so i think i was a bit over the top in writing as though speaking to a friend. Superlatively cleansed :)

And "website" and any "site"s have been removed
