### 2019-09-29 #53 @labarba
👋 @arfon --- something seems to be going wrong here. Help!
The DOI does redirect to `https://jose.theoj.org/papers/10.21105/jose.00053` but this URL gives `404 Not Found`. This is the first JOSE paper published since the site upgrade, so maybe a bug on our side?
