### 2019-09-29 #53 @arfon
> This is the first JOSE paper published since the site upgrade, so maybe a bug on our side?

Yeah, I think it is but unfortunately it might be tricky to chase down. Can you ping me before you next accept a JOSE paper so I can watch the logs?

For now, I've fixed up this paper.
