### 2019-09-29 #53 @labarba
Congratulations, @AstrobioMike, your JOSE paper is now published! 🚀

Many many thanks to our reviewers: @luizirber, @JasonJWilliamsNY — we couldn't do this without you! 🙏
